Fix(webhooks.py): Fixes JSON parsing errors in Python 3
This commit fixes some errors caused by changes to Flask and the JSON
parsing libraries in Python 3.

The change to .get_json() fixes an error because the Flask request.data
is encoded as a bytestring in python3, and loads() only supports string
as a datatype. `loads(request.data.decode('utf-8'))` would be the
equivalent.

Decoded the `stdout` and `stderr` to fix parsing errors with `dumps()`